changed movement of side devices
