+++ Assets/GameObjects/home_assistant.prefab
@@ -465,6 +465,7 @@ GameObject:
   - component: {fileID: 3463239637282198608}
   - component: {fileID: 3351004518092303875}
   - component: {fileID: 8760343801841549423}
+  - component: {fileID: 7082745337369395699}
   m_Layer: 0
   m_Name: home_assistant
   m_TagString: Untagged
@@ -572,6 +573,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   myChildObject: {fileID: 0}
+--- !u!114 &7082745337369395699
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6305877045556163205}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 324e3e04434d88c49ae44da80a0ca3ec, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &6370530675980487285
 GameObject:
   m_ObjectHideFlags: 0
@@ -763,7 +776,6 @@ GameObject:
   - component: {fileID: 1088838285769352696}
   - component: {fileID: 2458642877605765248}
   - component: {fileID: 4776917484347795051}
-  - component: {fileID: 5268615174004598061}
   m_Layer: 5
   m_Name: View Score
   m_TagString: Untagged
@@ -899,21 +911,6 @@ MonoBehaviour:
   empty: {fileID: 21300000, guid: b1e08492365043c4a9922368aa1fbd69, type: 3}
   ethernet: {fileID: 21300000, guid: bb793e407ef71704bb3208105ad63a93, type: 3}
   score: {fileID: 404360688598763986}
---- !u!114 &5268615174004598061
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7267048382567514977}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6445b95c02342424eba4bb68586624f7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  empty: {fileID: 21300000, guid: b1e08492365043c4a9922368aa1fbd69, type: 3}
-  score: {fileID: 404360688598763986}
-  isShowing: 0
 --- !u!1 &7649354018304191830
 GameObject:
   m_ObjectHideFlags: 0
