+++ Assets/GameObjects/smartphone.prefab
@@ -298,7 +298,6 @@ GameObject:
   - component: {fileID: 8561950781239669554}
   - component: {fileID: 2385456667709522573}
   - component: {fileID: 4475045032090770413}
-  - component: {fileID: 4938486072027072860}
   m_Layer: 5
   m_Name: View Score
   m_TagString: Untagged
@@ -434,21 +433,6 @@ MonoBehaviour:
   empty: {fileID: 21300000, guid: b1e08492365043c4a9922368aa1fbd69, type: 3}
   ethernet: {fileID: 21300000, guid: bb793e407ef71704bb3208105ad63a93, type: 3}
   score: {fileID: 3727151928966048139}
---- !u!114 &4938486072027072860
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2719031289367311143}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6445b95c02342424eba4bb68586624f7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  empty: {fileID: 21300000, guid: b1e08492365043c4a9922368aa1fbd69, type: 3}
-  score: {fileID: 3727151928966048139}
-  isShowing: 0
 --- !u!1 &4178867639077822427
 GameObject:
   m_ObjectHideFlags: 0
@@ -1000,6 +984,7 @@ GameObject:
   - component: {fileID: 2556865783745449462}
   - component: {fileID: 6011410177587792257}
   - component: {fileID: 5643320710410703516}
+  - component: {fileID: 2703976653238533454}
   m_Layer: 0
   m_Name: smartphone
   m_TagString: Untagged
@@ -1095,3 +1080,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   myChildObject: {fileID: 0}
+--- !u!114 &2703976653238533454
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7464038335503800715}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 324e3e04434d88c49ae44da80a0ca3ec, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
