+++ Assets/Scripts/moveDevices.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 324e3e04434d88c49ae44da80a0ca3ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
